Passsword Rest

+++ src/controllers/productController.js
@@ -23,7 +23,17 @@ class productController {
     }
 
     async deleteProduct(pid) {
-        return await this.productService.deleteProduct(pid);
+        const user = req.session.user; // from viewsRouter.js
+        const product = await this.productService.getProductById(pid);
+        if(!product) {
+            throw new Error("Product not found");
+        }
+        if(user.role === "premium" && product.owner !== user.email) {
+            throw new Error("You are not allowed to delete this product");
+        }
+        await this.productService.deleteProduct(pid);
+
+        return product;
     }
 }
